+++ src/Settings/default_bottom.lua
@@ -4,9 +4,11 @@ _G.DEFAULT_LAYOUT_BOTTOM =
         ["vitals"] = {
             ["effects"] = {
                 ["debuffs"] = {
+                    ["alignment"] = 2.000000,
                     ["slot"] = 2.000000,
                 },
                 ["buffs"] = {
+                    ["alignment"] = 2.000000,
                     ["slot"] = 1.000000,
                 },
             },
@@ -19,9 +21,11 @@ _G.DEFAULT_LAYOUT_BOTTOM =
         ["vitals"] = {
             ["effects"] = {
                 ["debuffs"] = {
+                    ["alignment"] = 2.000000,
                     ["slot"] = 2.000000,
                 },
                 ["buffs"] = {
+                    ["alignment"] = 2.000000,
                     ["slot"] = 1.000000,
                 },
             },